requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
